Rename regional_disruptions_overview to regional_disruption_impact_overview
Just renaming to clarify this is the impact query. This is a really good query.

Making the graph for this will be tough, especially the heatmap. We'll probably need to modify this in the future to filter by event type or something. Heatmap may require separating out impact counts depending on how we design it, so this query is great for that